+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * `freebsd_package` resource now uses the brand new "pkgng" package manager when available. (CHEF-4637)
 * chef-full template gets knife options to override install script url, add wget/curl cli options, and custom install commands (CHEF-4697)
 * knife now bootstraps node with the latest current version of chef-client. (CHEF-4911)
+* Add a threaded download queue for synchronizing cookbooks in parallel. (CHEF-4423)
 
 ## Last Release: 11.12.0 RC1 (03/31/2014)
 * SIGTERM will once-more kill a non-daemonized chef-client (CHEF-5172)